backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "FastAPI Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlmodel",
    "alembic",
    "psycopg2-binary",
    "psycopg2-binary",
    "asyncpg",
    "pyjwt",
    "httpx",
    "cryptography",
    "python-dotenv",
    "requests"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
